Move proxy errors to proxy/common/errors
